06_Streaming/requirements.txt
PyYAML
pyspark
kafka==1.3.5
